Refactor core modules
