--- src/controlflow/core/agent/__init__.py
@@ -1,2 +1,0 @@
-from . import memory
-from .agent import Agent, get_default_agent
