+++ src/controlflow/decorators.py
@@ -3,9 +3,9 @@ import inspect
 from typing import Any, Callable, Optional, Union
 
 import controlflow
-from controlflow.core.agent import Agent
-from controlflow.core.flow import Flow
-from controlflow.core.task import Task
+from controlflow.agents import Agent
+from controlflow.flows import Flow
+from controlflow.tasks.task import Task
 from controlflow.utilities.logging import get_logger
 from controlflow.utilities.prefect import flow as prefect_flow
 from controlflow.utilities.prefect import task as prefect_task
