+++ src/controlflow/llm/__init__.py
@@ -1,2 +1,1 @@
-from ..core.flow import history
 from . import models, messages, tools, handlers, completions
