+++ tests/ai_tests/test_tasks.py
@@ -1,4 +1,3 @@
-import pandas as pd
 import pytest
 from controlflow import Task
 from pydantic import BaseModel
@@ -20,13 +19,3 @@ class TestTaskResults:
         result = task.run()
         assert isinstance(result, Name)
         assert result == Name(first="John", last="Doe")
-
-    @pytest.mark.xfail(reason="Need to revisit dataframe handling")
-    def test_task_dataframe_result(self):
-        task = Task(
-            'return a dataframe with column "x" that has values 1 and 2 and column "y" that has values 3 and 4',
-            result_type=pd.DataFrame,
-        )
-        result = task.run()
-        assert isinstance(result, pd.DataFrame)
-        assert result.equals(pd.DataFrame(data={"x": [1, 2], "y": [3, 4]}))
