+++ tests/fixtures/__init__.py
@@ -1,4 +1,2 @@
-from .mocks import *
 from .instructions import *
-
 from .controlflow import *
